Refactor memory management and architecture implementation
- Removed unused imports from `mapping_area.rs`, `space.rs`, and `mod.rs`.
- Introduced a new `arch_impl.rs` file to provide common architecture implementations for `VirtualMemory` and `Arch`.
- Added `memory_impl.rs` to define `UserAddressSpace` and `KernAddressSpace` implementations for different architectures.
- Created mock architecture files under `mock/` directory to facilitate testing when not targeting RISC-V or LoongArch.
- Implemented basic mock functions for boot, interrupts, timers, and traps to support architecture-agnostic code compilation.
- Defined Linux syscall numbers in `syscall.rs` for consistent ABI across architectures.
