+++ os/src/arch/loongarch/intr/mod.rs
@@ -1,23 +1,15 @@
 //! LoongArch64 中断处理模块
-#![allow(unused)]
+#![allow(unsafe_op_in_unsafe_fn)]
 
-use crate::{
-    arch::constant::{CSR_CRMD_IE, CSR_ECFG_LIE_MASK, SSTATUS_SIE},
-    earlyprintln,
-};
-
-// CSR 编号
-const CRMD_CSR: u32 = 0x0;
-const ECFG_CSR: u32 = 0x4;
+use crate::arch::constant::{CSR_CRMD_IE, CSR_ECFG_LIE_MASK, SSTATUS_SIE};
 
 // 本地中断位
 const TIMER_LIE_BIT: usize = 1 << 11; // LIT（Local Interrupt Timer）对应的使能位
 
 #[inline(always)]
 unsafe fn set_crmd_ie(enable: bool) -> usize {
-    // 读取 CRMD，修改 IE 位后回写，返回旧值
     let old: usize;
-    let mut new: usize;
+    let new: usize;
     core::arch::asm!(
         "csrrd {old}, 0x0",
         old = out(reg) old,
@@ -51,7 +43,6 @@ unsafe fn read_crmd() -> usize {
 
 #[inline(always)]
 unsafe fn update_ecfg(mask: usize, set: bool) {
-    // 只改动给定 mask 覆盖的位
     let mut val: usize;
     unsafe {
         core::arch::asm!(
@@ -79,15 +70,7 @@ unsafe fn update_ecfg(mask: usize, set: bool) {
 /// # Safety
 /// 直接操作 CSR，调用者需确保时序正确
 pub unsafe fn enable_timer_interrupt() {
-    earlyprintln!("Enabling timer interrupt");
     unsafe { update_ecfg(TIMER_LIE_BIT, true) };
-}
-
-/// 禁用定时器中断（仅清除本地定时器使能位）
-/// # Safety
-/// 直接操作 CSR，调用者需确保时序正确
-pub unsafe fn disable_timer_interrupt() {
-    unsafe { update_ecfg(TIMER_LIE_BIT, false) };
 }
 
 /// 启用全局中断
@@ -117,20 +100,15 @@ pub fn are_interrupts_enabled() -> bool {
 /// 读取并禁用中断（返回之前的 CRMD 值）
 /// # Safety
 /// 直接操作 CSR 寄存器
+#[allow(dead_code)]
 pub unsafe fn read_and_disable_interrupts() -> usize {
     unsafe { set_crmd_ie(false) }
-}
-
-/// 读取并启用中断（返回之前的 CRMD 值）
-/// # Safety
-/// 直接操作 CSR 寄存器
-pub unsafe fn read_and_enable_interrupts() -> usize {
-    unsafe { set_crmd_ie(true) }
 }
 
 /// 恢复中断状态
 /// # Safety
 /// 直接操作 CSR 寄存器
+#[allow(dead_code)]
 pub unsafe fn restore_interrupts(flags: usize) {
     if flags & SSTATUS_SIE != 0 {
         unsafe { enable_interrupts() };
@@ -140,11 +118,13 @@ pub unsafe fn restore_interrupts(flags: usize) {
 }
 
 /// 启用指定 IRQ（LoongArch 目前仅处理本地中断位，其他由中断控制器驱动负责）
+#[allow(dead_code)]
 pub fn enable_irq(_irq: usize) {
     // 留给外部中断控制器驱动实现
 }
 
 /// 禁用指定 IRQ
+#[allow(dead_code)]
 pub fn disable_irq(_irq: usize) {
     // 留给外部中断控制器驱动实现
 }
@@ -152,5 +132,6 @@ pub fn disable_irq(_irq: usize) {
 /// 软中断模块
 pub mod softirq {
     /// 初始化软中断
+    #[allow(dead_code)]
     pub fn init() {}
 }
