+++ os/src/arch/mock/boot.rs
@@ -1,0 +1,5 @@
+pub fn main(_hartid: usize) -> ! {
+    loop {
+        core::hint::spin_loop();
+    }
+}
