+++ os/src/arch/mock/constant.rs
@@ -1,0 +1,6 @@
+pub const SSTATUS_SIE: usize = 1;
+pub const ARCH: &str = "mock";
+pub const USER_TOP: usize = 0x7fff_ffff_ffff;
+pub const SUPERVISOR_EXTERNAL: usize = 0;
+pub const SV39_BOT_HALF_TOP: usize = 0x8000_0000_0000;
+pub const STACK_ALIGN_MASK: usize = !0xf;
