+++ os/src/arch/mock/ipi.rs
@@ -1,0 +1,5 @@
+pub fn send_reschedule_ipi(_cpu: usize) {}
+
+pub fn send_tlb_flush_ipi_all() {}
+
+pub fn handle_ipi() {}
